Training: reject fp16 full-FT + GradScaler; default precision to bf16
- Full fine-tune with weight_dtype=fp16 + training_dtype=fp16 +
  mixed_precision crashed at step 1 ('Attempting to unscale FP16
  gradients') - GradScaler.unscale_ requires FP32 master params, and
  full-FT trains the fp16 base weights directly. LoRA is unaffected
  (adapters are fp32). Now rejected at setup with an actionable error
- The crash combo was the OUT-OF-BOX default: weight_dtype /
  training_dtype defaults flipped fp16 -> bf16 in param_defaults.py
  (single source of truth) + openapi.yaml (bf16 needs no grad scaling
  and works for both LoRA and full-FT)

+++ backend/api/param_defaults.py
@@ -240,8 +240,8 @@ TRAINING_DEFAULTS: Dict[str, Any] = {
     "force_recache": False,
     "reconstruction_loss_weight": 0.0,
     # Precision
-    "weight_dtype": "fp16",
-    "training_dtype": "fp16",
+    "weight_dtype": "bf16",                 # bf16: works for both LoRA and full-FT
+    "training_dtype": "bf16",               # bf16 needs no GradScaler (fp16 full-FT crashes on unscale_)
     "output_dtype": "fp32",
     "vae_dtype": "fp16",                    # Fix: frontend had "fp32"
     # Full-parameter save: also embed the VAE weights into the single-file
